requirements.txt
black==23.3.0
emoji==2.9.0
errant==2.3.3
fastbm25==0.0.2
gpustat==1.1
huggingface_hub==0.15.1
nltk==3.8.1
omegaconf==2.3.0
OpenCC==1.1.6
pathos==0.3.0
python-Levenshtein==0.21.1
sacrebleu==2.4.2
scipy==1.10.1
torch==2.0.1
transformers==4.29.2